html/rendering/bindings/the-keygen-element-0/htmlkeygenelement_form.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: htmlkeygenelement_form</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/forms.html#dom-fae-form">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div>
      <form id="f">
        <keygen id="security1"/>
      </form>
      <keygen id="security2"/>
    </div>
    <script>
      test(function() {
        assert_equals(document.getElementById("security1").form, document.getElementById("f"), "Has a form owner.");
      }, "htmlkeygenelement_form_base");

      test(function() {
        assert_equals(document.getElementById("security2").form, null, "Has not a form owner.");
      }, "htmlkeygenelement_no_owner");
    </script>
  </body>
</html>
